non-RME methods from RMEMatcher moved into SingleFeatureMultiCrop
